pyproject.toml
[project]
name = "floodguard"
version = "0.1.0"
description = "Punjab Flood Monitoring and Simulation Web Application"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "folium>=0.20.0",
    "geoalchemy2>=0.18.0",
    "geojson>=3.2.0",
    "geopandas>=1.1.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.0",
    "rasterio>=1.4.3",
    "shapely>=2.1.1",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
# Explicitly tell setuptools what to include and exclude
py-modules = ["main", "api_routes", "models", "seed_data", "mock_data", "database"]

[tool.setuptools.packages.find]
# Don't try to find packages automatically
where = ["."]
exclude = ["frontend*", "attached_assets*", "tests*", ".github*"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
